+++ src/conditioned_kernel/return_path/__init__.py
@@ -3,7 +3,7 @@
 from conditioned_kernel.return_path.accept import accept_candidate
 from conditioned_kernel.return_path.assess import assess
 from conditioned_kernel.return_path.parse import parse_candidate
-from conditioned_kernel.return_path.repair import build_repair_annotations
+from conditioned_kernel.return_path.repair import build_repair_annotations, build_repair_plan
 from conditioned_kernel.return_path.validate import validate_candidate
 
 __all__ = [
@@ -11,5 +11,6 @@ __all__ = [
     "validate_candidate",
     "assess",
     "build_repair_annotations",
+    "build_repair_plan",
     "accept_candidate",
 ]